% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous-ChineseColor.R
\name{pal_ChineseColor}
\alias{pal_ChineseColor}
\title{Different types of Chinese Color palettes grouped by basic colors.}
\usage{
pal_ChineseColor(
  palette = c("blue", "black", "cyan", "gray", "green", "purple", "red", "white",
    "yellow"),
  n = 3,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{n}{Number of individual colors to be generated.}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{Logical. Should the order of the colors be reversed?}
}
\description{
Mimic from ggsci(https://github.com/nanxstats/ggsci)
}
\examples{
pal_ChineseColor("cyan")
pal_ChineseColor("green", n = 30, alpha = 0.6, reverse = TRUE)
}
\references{
\url{https://colors.masantu.com/#/}
}
\author{
Yu Zhong Peng <\email{mugpeng@foxmail.com}>
}
